Address more feedback from PR.
